Add Python 3.10 jobs on Fedora 35
Since there are no Python 3.10 jobs upstream in zuul-jobs, we have to
inherit from the generic `tox` job here.

Switch what we can to 3.10/f35, but keep all the older jobs existing
because this is a three-step process in Zuul: first we define new jobs,
then we switch the downstream repo to the new jobs, and only than the
old jobs can be removed.

Change-Id: Ia9fdf0004e331291aa5c31f27b55f0fe32572a7c